requirements.txt
absl-py==0.14.1
anyio==3.3.3
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cycler==0.10.0
docopt==0.6.2
fastapi==0.70.0
gast>=0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h11==0.12.0
h5py>=2.10.0,<3.2
idna==3.2
joblib==1.1.0
Keras>=2.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.3.4
numpy>=1.7,<1.23.0
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.3.2
pipreqs==0.4.10
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
six>=1.5,<1.16.0
sniffio==1.2.0
starlette==0.16.0
tensorboard==2.6
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.2
typing-extensions>=3.7.4
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
wrapt==1.12.1
yarg==0.1.9
